Try to introduce an enum of standard well-known build systems
This doesn't work as well as I'd hoped. The main problem is that there's now
no obvious place to define the ID contants for Gradle and Maven as they can't
be defined on StandardBuildSystem which has replaced GradleBuildSystem and
MavenBuildSystem. An alternative would be to add a StandardBuildSystem
attribute to ConditionalOnBuildSystem alongside the existing String attribute
which would be renamed to id. Unfortunately, an enum attribute cannot have
a default value of null which is what we'd want it to be.