Further polish help document generation
This commit removes the duplicate link representation in favour of
a contributor that reuses the metadata.

HelpDocument has been further polished as well with a dedicated concept
for Getting Started and dedicated methods to register link using the
individual elements as done in other part of the codebase.

See gh-38