Remove concept of "model"
The separate package model was added as a temporary measure for concepts
coming from the metadata but this was a mistake. It isn't very clear to
me just yet how existing concepts will map to concepts in the new API
but there is an interest to make them as separate as possible.

While this is more work for contributors, new APIs are working with a
narrowed view of the model and prevent access to other parts of the
model.