Rename dependencies to requestedDependencies
This commit makes it clearer that the dependencies set in
ProjectDescription is about the dependencies requested by the user and
not the ultimate set that is used to generate the build configuration.

ConditionalOnDependency has been renamed consistently and it is now
using the id of a dependency rather than the full coordinates that can
change according to the platform version.

Closes gh-60
Closes gh-41